Move mock Github test classes for future re-use.
